Rename DADGenerator to DataAccessDocGenerator
